% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.hgld.R
\name{diag.hgld}
\alias{diag.hgld}
\title{Diagnostic for the Hurdle Generalized Lambda Distribution fit}
\usage{
diag.hgld(fit, facet = FALSE, facet.breaks, facet.labels, facet.ncol,
  trace = TRUE, no.test = 1000, len = floor(0.9 * length(fit$data[fit$data
  != 0])), alpha = 0.05, plotKS = TRUE, KS = TRUE)
}
\arguments{
\item{fit}{An object of class \link[HGLD]{fit.hgld}.}

\item{facet}{Whether the plots must be faceted for better visualization.}

\item{facet.breaks}{The breaks in which to facet the data. Must be the endpoints of the intervals.}

\item{facet.labels}{The labels of the categories given by the facet breaks.}

\item{facet.ncol}{The number of columns for the facet plot.}

\item{trace}{Whether progress bar must be printed in order to trace the algorithm.}

\item{no.test}{Total number of KS tests required.}

\item{len}{Number of data to sample at each KS test.}

\item{alpha}{Significance level of KS test.}

\item{plotKS}{Whether to plot the KS resample test result within each plot.}

\item{KS}{Whether the resample KS test must be performed to the nonzero values.}
}
\value{
\item{qqRS}{\link[ggplot2]{ggplot} qq-plot for the RS HGLD.}

\item{qqFMKL}{\link[ggplot2]{ggplot} qq-plot for the FMKL HGLD.}

\item{quantRS}{\link[ggplot2]{ggplot} quantile plot for the RS HGLD.}

\item{quantFMKL}{\link[ggplot2]{ggplot} quantile plot for the FMKL HGLD.}

\item{moments}{Moments comparison for the GLD fitted to the nonzero data for both parametrizations. Those are not the
moments of the HGLD.}

\item{KS}{Percentage of no rejection for the KS resample test.}
}
\description{
Diagnostic plots and measures of goodness-of-fit for a Hurdle Generalized Lambda Distribution fit.
}
\details{
The diagnostics techniques are applied to the nonzero data values. Returns the qq-plot and the quantile plot between the
data and the theoretical fitted HGLD. Also returns a table comparing sample moments with theoretical moments. A Kolmogorov-Simornov resample
test is performed and the percentage of the times that the null hypotheses, i.e., goodness-of-fit, is not rejected is displayed.
All diagnostics are performed for both the RS and FMKL GLD.
}
\examples{
set.seed(100)
data <- healthcare[sample(1:nrow(healthcare),100),]
fit <- fit.hgld(data$log_expense)
d <- suppressWarnings(diag.hgld(fit,facet = FALSE,plotKS = FALSE))

#mixture
set.seed(100)
data <- c(rcauchy(50,location = 10),rep(0,30),rcauchy(50))
fit <- fit.hgld(data = data,mixture = TRUE)
d <- suppressWarnings(diag.hgld(fit))

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; Fitting a Hurdle Generalized Lambda Distribution to healthcare expenses. (2017) \emph{arxiv1712.02183}

Su, S. Fitting Single and Mixture of Generalized Lambda Distributions to Data via Discretized and Maximum Likelihood Methods: GLDEX in R. (2007), Journal of Statistical Software: *21* 9.
}
