% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.hgld.R
\name{fit.hgld}
\alias{fit.hgld}
\title{Fit the Hurdle Generalized Lambda Distribution}
\usage{
fit.hgld(data, mixture = FALSE, clustering.m = clara, threshold = NULL,
  leap1 = 3, leap2 = 3, fun1 = "runif.sobol", fun2 = "runif.sobol",
  rs.leap = 3, fmkl.leap = 3, rs.init = c(-1.5, 1.5),
  fmkl.init = c(-0.25, 1.5), FUN = "runif.sobol", no = 10000)
}
\arguments{
\item{data}{A vector of data.}

\item{mixture}{Whether a mixture of HGLD must be fitted.}

\item{clustering.m}{Clustering method used in classifying the dataset into two parts when fitting a mixture of HGLD.
Valid arguments include clara, fanny and pam from the cluster library, or threshold, if the data must be divided
by a threshold. Default is clara. Or a logical vector specifying how data should be split. See
\link[GLDEX]{fun.auto.bimodal.pml} for more details.}

\item{threshold}{The threshold to divide the data if \emph{clustering.m = "threshold"}.}

\item{leap1}{Scrambling (0,1,2,3) for the sobol sequence for the first distribution fit when fitting a mixture of
HGLD. See scrambling/leap argument for \link[GLDEX]{runif.sobol}, \link[GLDEX]{runif.halton} or
\link[GLDEX]{QUnif} of the \link[GLDEX]{GLDEX} package.}

\item{leap2}{Scrambling (0,1,2,3) for the sobol sequence for the second distribution fit when fitting a mixture of
HGLD. See scrambling/leap argument for \link[GLDEX]{runif.sobol}, \link[GLDEX]{runif.halton} or
\link[GLDEX]{QUnif} of the \link[GLDEX]{GLDEX} package.}

\item{fun1}{A character string of either "runif.sobol" (default), "runif.halton" or "QUnif" for the first
distribution fit when fitting a mixture of HGLD. See \link[GLDEX]{fun.auto.bimodal.pml} for more details.}

\item{fun2}{A character string of either "runif.sobol" (default), "runif.halton" or "QUnif" for the second
distribution fit when fitting a mixture of HGLD. See \link[GLDEX]{fun.auto.bimodal.pml} for more details.}

\item{rs.leap}{Scrambling (0,1,2,3) for the sobol sequence for the RS generalized lambda distribution fit. See scrambling/leap
argument for \link[GLDEX]{runif.sobol}, \link[GLDEX]{runif.halton} or \link[GLDEX]{QUnif} of the
\link[GLDEX]{GLDEX} package. See \link[GLDEX]{fun.data.fit.ml} for more details.}

\item{fmkl.leap}{Scrambling (0,1,2,3) for the sobol sequence for the FMKL generalized lambda distribution fit. See scrambling/leap
argument for \link[GLDEX]{runif.sobol}, \link[GLDEX]{runif.halton} or \link[GLDEX]{QUnif} of the
\link[GLDEX]{GLDEX} package. See \link[GLDEX]{fun.data.fit.ml} for more details.}

\item{rs.init}{Initial values (lambda3 and lambda4) for the RS generalized lambda distribution. See
\link[GLDEX]{fun.data.fit.ml} for more details.}

\item{fmkl.init}{Initial values (lambda3 and lambda4) for the FMKL generalized lambda distribution. See
\link[GLDEX]{fun.data.fit.ml} for more details.}

\item{FUN}{A character string of either "runif.sobol" (default), "runif.halton" or "QUnif". See
\link[GLDEX]{fun.data.fit.ml} for more details.}

\item{no}{Number of initial random values to find the best initial values for optimization. See
\link[GLDEX]{fun.data.fit.ml} for more details.}
}
\value{
\item{par}{The estimate of the HGLD five parameters for both the RS and FMKL parametrizations if \emph{mixture
= FALSE}. Otherwise present the estimation of ten parameters: the zero probability mass, the four parameters of
each GLD and the clustering parameter p.}

\item{data}{The data used in the fit.}

\item{mixture}{Whether a mixture of HGLD was fitted.}
}
\description{
Fit the Hurdle Generalized Lambda Distribution to a dataset by numerical maximum likelihood.
}
\details{
Given a dataset, estimate by the numerical maximum likelihood method the five parameters of the HGLD.
Fit both the RS and the FMKL parametrizations. Also fit mixture of HGLD.
}
\examples{
set.seed(100)
data <- healthcare[sample(1:nrow(healthcare),100),]
fit <- fit.hgld(data$log_expense)

#mixture
set.seed(100)
data <- c(rcauchy(50,location = 10),rep(0,30),rcauchy(50))
fit <- fit.hgld(data = data,mixture = TRUE)

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; A Survey of a Hurdle Model for Heavy-Tailed Data Based on the Generalized Lambda Distribution. (2017) \emph{arxiv1712.02183}

Su, S.; Fitting Single and Mixture of Generalized Lambda Distributions to Data via Discretized and Maximum Likelihood Methods: GLDEX in R.(2007), Journal of Statistical Software: *21* 9.
}
