% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fit.hgld.R
\name{plot.fit.hgld}
\alias{plot.fit.hgld}
\title{Plot the Hurdle Generalized Lambda Distribution fit}
\usage{
\method{plot}{fit.hgld}(x, histogram = TRUE, hcolor = "white",
  hfill = "black", bins = 50, dcolor = c("black", "red"),
  dtype = c("solid", "solid"), xlab = "Data", ylab = "Density", ...)
}
\arguments{
\item{x}{A \link[HGLD]{fit.hgld} object.}

\item{histogram}{Logical. Whether the estimated density must be superimposed by the data histogram.}

\item{hcolor}{Color of the histogram.}

\item{hfill}{Color to fill the histogram.}

\item{bins}{Number of histogram bins.}

\item{dcolor}{Color of the RS and FMKL density plot.}

\item{dtype}{Type of the RS and FMKL density line.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-axis.}

\item{...}{Arguments to be passed to methods.}
}
\value{
\item{RS}{\link[ggplot2]{ggplot} plot of the fitted RS GLD.}

\item{FMKL}{\link[ggplot2]{ggplot} plot of the fitted FMKL GLD.}

\item{RF}{\link[ggplot2]{ggplot} plot of the fitted RS and FMKL GLD on the same plot.}
}
\description{
Density plot of the fit given by an object of class \link[HGLD]{fit.hgld}.
}
\details{
The density may be plotted by itself or superimposed by the data histogram. Plot only the nonzero data values.
}
\examples{
set.seed(100)
data <- healthcare[sample(1:nrow(healthcare),100),]
fit <- fit.hgld(data$log_expense)
plot(fit)

#mixture
set.seed(100)
data <- c(rcauchy(50,location = 10),rep(0,30),rcauchy(50))
fit <- fit.hgld(data = data,mixture = TRUE)
plot(fit)

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; Fitting a Hurdle Generalized Lambda Distribution to healthcare expenses. (2017) \emph{arxiv1712.02183}
}
