% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reg.hgld.R
\name{plot.reg.hgld}
\alias{plot.reg.hgld}
\title{Predict density plot of a Hurdle Generalized Lambda Regression}
\usage{
\method{plot}{reg.hgld}(x, newvalues, name = row.names(newvalues),
  color = NULL, xlab = "Data",
  xlim = c(min(x$NZdata[[all.vars(x$loc.formula)[1]]]),
  max(x$NZdata[[all.vars(x$loc.formula)[1]]])), title = "", ...)
}
\arguments{
\item{x}{An object of class \link[HGLD]{reg.hgld}.}

\item{newvalues}{A data frame with the new values of the covariates. Column names must match the ones given in formulas
\emph{loc.formula} and \emph{zero.formula}.}

\item{name}{A vector with the names of the new values profiles. Default is the row names of the newvalues data frame.}

\item{color}{The color of each density curve. Must have one color for each row of newvalues.}

\item{xlab}{Label of the x-axis.}

\item{xlim}{A vector with the limits of the x-axis.}

\item{title}{Legend title.}

\item{...}{Arguments to be passed to methods.}
}
\value{
\item{plot}{\link[ggplot2]{ggplot} density plot for the given new values}
}
\description{
Predict density plot of a Hurdle Generalized Lambda Regression.
}
\details{
Given an object of class \link[HGLD]{reg.hgld} and new values for the covariates,
 returns the density function of the fitted HGLD, given the covariates. The contrast on the
 \emph{newvalues} data frame must be same of the \emph{data} used in the \link[HGLD]{reg.hgld} object. All
 density curves are plotted in the same plot.
}
\examples{
set.seed(100)
tmp <- na.omit(healthcare)
data <- tmp[sample(1:nrow(tmp),50),]
formula <- log_expense ~ age + sex + log_previous_expense
reg <- suppressWarnings(reg.hgld(data = data,zero.formula = formula,loc.formula = formula,
                                 full = FALSE,param = "rs"))
newvalues <- tmp[sample(1:nrow(tmp),5),c(2,3,5,8)]
plot(reg,newvalues)

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; A Survey of a Hurdle Model for Heavy-Tailed Data Based on the Generalized Lambda Distributions. (2017) \emph{arxiv1712.02183}
}
