% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhgld.R
\name{qhgld}
\alias{qhgld}
\title{The Hurdle Generalized Lambda Distribution Family}
\usage{
qhgld(p, mixture = FALSE, lambda1, lambda2 = NULL, prob = NULL,
  param = "fmkl", trace = FALSE, inverse.eps = 1e-08,
  max.iterations = 500)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{mixture}{Whether to give the density of a mixture HGLD.}

\item{lambda1}{A vector of length 5 with the five parameters of the HGLD or the first HGLD if mixture = TRUE.}

\item{lambda2}{A vector of length 4 with the four parameters of the second HGLD.}

\item{prob}{The cluster parameter for the mixture HGLD.}

\item{param}{"fmkl" or "rs".}

\item{trace}{Whether progress bar must be printed in order to trace the algorithm.}

\item{inverse.eps}{Accuracy of calculation for the numerical determination of F(x), defaults to 1e-8.}

\item{max.iterations}{Maximum number of iterations in the numerical determination of F(x), defaults to 500.}
}
\value{
The quantile function of a HGLD.
}
\description{
Quantile function of a Hurdle Generalized Lambda Distribution.
}
\details{
The HGLD must be non-negative, otherwise the quantile is only approximate. If the parametrization of
the RS or FMKL GLD is not acceptable, the function returns NA. This function is based on the \link[GLDEX]{GLDEX} package.
}
\examples{
set.seed(100)
data <- healthcare[sample(1:nrow(healthcare),100),]
fit <- fit.hgld(data$log_expense)
qhgld(p = seq(0,1,0.05),lambda1 = fit$par$RS,param = "rs")

#mixture
qhgld(p = seq(0,1,0.1),mixture = TRUE,lambda1 = c(0.05,0,1,3,6),lambda2 = c(0.3,2,3,6),prob = 0.5,
            param = "fmkl")

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; Fitting a Hurdle Generalized Lambda Distribution to healthcare expenses. (2017) \emph{arxiv1712.02183}

Su, S.; Fitting Single and Mixture of Generalized Lambda Distributions to Data via Discretized and Maximum Likelihood Methods: GLDEX in R. (2007), Journal of Statistical Software: *21* 9.
}
