% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare.R
\docType{data}
\name{healthcare}
\alias{healthcare}
\title{Healthcare data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 289456 rows and 8 columns.}
\usage{
healthcare
}
\value{
\item{ID}{The ID of the customer.}

\item{sex}{The sex of the customer.}

\item{age}{The age of the customer on the considered year.}

\item{expense}{The healthcare expense of the customer on the considered year.}

\item{log_expense}{The logarithm of the healthcare expense of the customer on the considered year.}

\item{year}{The considered year.}

\item{previous_expense}{The healthcare expense of the customer on the previous year.}

\item{log_previous_expense}{The logarithm of the healthcare expense of the customer on the previous year.}
}
\description{
Data set containing the healthcare expense of 129,257 customers of a Brazilian healthcare
company between 2006 and 2009.
}
\details{
The expenses are in Reais (Brazilian currency) and were deflated to the January 2006 value. In order to
fit the models of the paper in the references it is necessary to turn every observation whose expense is less than
R$ 100 into zero.
}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; Fitting a Hurdle Generalized Lambda Distribution to healthcare expenses. (2017) \emph{arxiv1712.02183}
}
\keyword{datasets}
