% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhgld.reg.R
\name{qhgld.reg}
\alias{qhgld.reg}
\title{Predict quantiles of a Hurdle Generalized Lambda regression}
\usage{
qhgld.reg(p, reg, newvalues, l0 = TRUE, location = TRUE)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{reg}{An object of class \link[HGLD]{reg.hgld}.}

\item{newvalues}{A data frame with the new values of the covariates. Column names must match the ones given in formulas \emph{loc.formula} and \emph{zero.formula}.}

\item{l0}{Whether to return the lambda0 of each profile.}

\item{location}{Whether to return the location of each profile.}
}
\value{
The quantile of the new values, according to the fitted HGLD regression.
}
\description{
Quantile function of a Hurdle Generalized Lambda Regression.
}
\details{
Given an object of class \link[HGLD]{reg.hgld}, a percentile and new values for the covariates,
 it returns the quantile of the fitted HGLD, given the covariates. The fitted HGLD must be non-negative. The
 contrast on the \emph{newvalues} data frame must be same of the \emph{data} used in the \link[HGLD]{reg.hgld} object.
}
\examples{
set.seed(10)
tmp <- na.omit(healthcare)
data <- tmp[sample(1:nrow(tmp),100),]
formula <- log_expense ~ age + sex + log_previous_expense
reg <- suppressWarnings(reg.hgld(data,formula,formula,TRUE,n.simu = 10,param = "rs",plotKS = TRUE))
newvalues <- tmp[sample(1:nrow(tmp),5),c(2,3,8)]
suppressWarnings(qhgld.reg(p = seq(0.05,0.95,0.1),reg,newvalues))

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; Fitting a Hurdle Generalized Lambda Distribution to healthcare expenses. (2017) \emph{arxiv1712.02183}
}
