% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.hgld.R
\name{reg.hgld}
\alias{reg.hgld}
\title{Fit a Hurdle Generalized Lambda Distribution Regression model}
\usage{
reg.hgld(data, zero.formula, loc.formula, full = FALSE, param = "fmkl",
  maxit = 20000, init = NULL, alpha = 0.05, n.simu = 1000,
  plotKS = TRUE, h.bins = 50)
}
\arguments{
\item{data}{A dataset containing the variables of the model.}

\item{zero.formula}{A symbolic expression of the model to be fitted to the hurdle parameter of the distribution.}

\item{loc.formula}{A symbolic expression of the model to be fitted to the location of the distribution.}

\item{full}{Whether a simulation method must be applied to derive a confidence interval for the location regression coefficients.}

\item{param}{"fmkl" or "rs".}

\item{maxit}{Maximum number of iterations for numerical optimization.}

\item{init}{Choose a different set of initial values to start the optimization process. This can either be full set of parameters including
GLD parameter estimates, or it can just be the coefficient estimates of the regression model.}

\item{alpha}{Significance level of the Confidence Interval for the GLD regression.}

\item{n.simu}{Number of times to repeat the simulation runs, defaults to 1000.}

\item{plotKS}{Whether to plot the KS resample test result within each plot.}

\item{h.bins}{Number of bins for the GLD Regression normalized quantiles residuals histogram.}
}
\value{
\item{coefficients}{The estimated coefficients of the HGLD regression.}

\item{Zplot}{A function that generates the four diagnostic plots of the logistic regression.}

\item{Zres.sumarry}{Summary of the normalised quantile residuals of the logistic regression.}

\item{Zfit}{Normalised quantile residuals versus fitted values for the logistic regression.}

\item{Zindex}{Normalised quantile residuals versus index for the logistic regression.}

\item{Zdensity}{Normalised quantile residuals density for the logistic regression.}

\item{Zqq}{QQ-norm plot of the normalised quantile residuals for the logistic regression.}

\item{NZqq}{QQ-plot for the GLD regression residuals.}

\item{NZquant}{Quantile plot for the GLD regression residuals.}

\item{NZhistogram}{The histogram of the GLD regression residuals.}

\item{NZplot}{The normalised quantile residuals plots for the GLD regression.}

\item{NZres.sumarry}{Summary of the normalised quantile residuals of the GLD regression.}

\item{NZfit}{Normalised quantile residuals versus fitted values for the GLD regression.}

\item{NZindex}{Normalised quantile residuals versus index for the GLD regression.}

\item{NZdensity}{normalised quantile residuals density for the GLD regression.}

\item{NZqqQuant}{QQ-norm plot of the normalised quantile residuals for the GLD regression.}

\item{KS}{KS test p-value for the GLD regression.}

\item{gamlss}{The \link[gamlss]{gamlss} object of the fitted logistic regression.}

\item{GLDreg}{The \link[GLDreg]{GLDreg} object of the fitted GLD regression.}

\item{Zdata}{The data used to fit the logistic regression}

\item{NZdata}{The data used to fit the GLD regression.}

\item{zero.formula}{A symbolic expression of the model to be fitted to the hurdle parameter of the distribution.}

\item{loc.formula}{A symbolic expression of the model to be fitted to the location of the distribution.}

\item{param}{"fmkl" or "rs".}

\item{full}{Whether a simulation method must be applied to derive a confidence interval for the location regression coefficients.}
}
\description{
Fit a Hurdle Generalized Lambda Distribution Regression model to a dataset.
}
\details{
Given a dataset, estimate by the Numerical Maximum Likelihood method the regression coefficients of the model and
the five parameters of the error GLD. The regression coefficients that model the location of the distribution are estimated by
the functions \link[GLDreg]{GLD.lm} and \link[GLDreg]{GLD.lm.full} of package \link[GLDreg]{GLDreg}. The regression coefficients
that model the hurdle parameter of the distribution are estimated by function \link[gamlss]{gamlss}.
}
\examples{
set.seed(100)
tmp <- na.omit(healthcare)
data <- tmp[sample(1:nrow(tmp),50),]
formula <- log_expense ~ age + sex + log_previous_expense
reg <- suppressWarnings(reg.hgld(data = data,zero.formula = formula,loc.formula = formula,
                                 full = FALSE,param = "rs"))

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; A Survey of a Hurdle Model for Heavy-Tailed Data Based on the Generalized Lambda Distribution. (2017) \emph{arxiv1712.02183}

Su, S.; Flexible Parametric Quantile Regression Model. (2015), Statistics & Computing May 2015, Volume 25, Issue 3, pp 635-650
}
