% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhgld.R
\name{rhgld}
\alias{rhgld}
\title{The Hurdle Generalized Lambda Distribution Family}
\usage{
rhgld(n, mixture = FALSE, lambda1, lambda2 = NULL, prob = NULL,
  param = "fmkl")
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{mixture}{Whether to give the density of a mixture HGLD.}

\item{lambda1}{A vector of length 5 with the five parameters of the HGLD or the first HGLD if mixture = TRUE.}

\item{lambda2}{A vector of length 4 with the four parameters of the second HGLD.}

\item{prob}{The cluster parameter for the mixture HGLD.}

\item{param}{"fmkl" or "rs".}
}
\value{
A sample of a HGLD.
}
\description{
Sample from a Hurdle Generalized Lambda Distribution.
}
\details{
If the parametrization of the RS or FMKL GLD is not acceptable, the function returns NA. This function is based on the \link[GLDEX]{GLDEX} package.
}
\examples{
library(ggplot2)
set.seed(100)
qplot(rhgld(n = 1000,lambda1 = c(0.05,0,1,3,6),param = "fmkl"),..density..,geom = "histogram",
      bins = 30)
qplot(rhgld(n = 1000,mixture = TRUE,lambda1 = c(0.05,0,1,3,6),lambda2 = c(0.3,2,3,6),prob = 0.5,
            param = "fmkl"),geom = "histogram",bins = 50)

}
\references{
Marcondes, D.; Peixoto, C.; Maia, A. C.; A Survey of a Hurdle Model for Heavy-Tailed Data Based on the Generalized Lambda Distribution. (2017) \emph{arxiv1712.02183}

Su, S.; Fitting Single and Mixture of Generalized Lambda Distributions to Data via Discretized and Maximum Likelihood Methods: GLDEX in R. (2007), Journal of Statistical Software: *21* 9.
}
